pyproject.toml
[project]
name = "reinnovation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-google-genai>=4.1.3",
    "langgraph>=1.0.5",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
